+++ packages/emails/.env.development
@@ -1,0 +1,1 @@
+PUBPUB_URL=http://localhost:3000
