+++ packages/emails/src/singup-invite.tsx
@@ -1,0 +1,90 @@
+import * as React from "react";
+import {
+	Body,
+	Button,
+	Column,
+	Container,
+	Head,
+	Heading,
+	Hr,
+	Html,
+	Img,
+	Link,
+	Preview,
+	Row,
+	Section,
+	Tailwind,
+	Text,
+} from "@react-email/components";
+
+import type { Communities, CommunitiesId } from "db/public";
+
+interface SignupInviteProps {
+	signupLink: string;
+	community: Communities;
+}
+
+const baseUrl = process.env.PUBPUB_URL ?? process.env.NEXT_PUBLIC_PUBPUB_URL ?? "";
+
+export const SignupInvite = ({
+	community = {
+		name: "CrocCroc",
+		id: "croccroc" as CommunitiesId,
+		slug: "croccroc",
+		avatar: `${baseUrl}/static/croccroc.png`,
+	},
+	signupLink,
+}: SignupInviteProps) => {
+	const previewText = `your PubPub password`;
+
+	return (
+		<Html>
+			<Head />
+			<Preview>{previewText}</Preview>
+			<Tailwind>
+				<Body className="mx-auto my-auto bg-white font-sans">
+					<Container className="mx-auto my-[40px] w-[465px] rounded border border-solid border-[#eaeaea] p-[20px]">
+						<Section className="mt-[32px]">
+							<Img
+								src={community.avatar ?? `${baseUrl}/static/logo.svg`}
+								width="40"
+								height="40"
+								alt="PubPub"
+								className="mx-auto my-0"
+							/>
+						</Section>
+						<Heading className="mx-0 my-[30px] p-0 text-center text-[24px] font-normal text-black">
+							Join {community.name} on PubPub
+						</Heading>
+						<Text className="text-[14px] leading-[24px] text-black">
+							You have been invited to signup for PubPub. Click the button below to
+							join {community.name} on PubPub.
+						</Text>
+						<Section className="mb-[32px] mt-[32px] text-center">
+							<Button
+								className="rounded bg-[#000000] px-4 py-3 text-center text-[12px] font-semibold text-white no-underline"
+								href={signupLink}
+							>
+								Join PubPub
+							</Button>
+						</Section>
+						<Text className="text-[14px] leading-[24px] text-black">
+							or copy and paste this URL into your browser:{" "}
+							<Link href={signupLink} className="text-blue-600 no-underline">
+								{signupLink}
+							</Link>
+						</Text>
+						<Hr className="mx-0 my-[26px] w-full border border-solid border-[#eaeaea]" />
+						<Text className="text-[12px] leading-[24px] text-[#666666]">
+							If you were not expecting this invitation, you can ignore this email. If
+							you are concerned about your account's safety, please reply to this
+							email to get in touch with us.
+						</Text>
+					</Container>
+				</Body>
+			</Tailwind>
+		</Html>
+	);
+};
+
+export default SignupInvite;
